requirements.txt
selenium>=4.0.0
webdriver-manager>=3.0.0
numpy>=2.2.0
Pillow>=10.0.0
opencv-python>=4.12.0.88
torch==2.9.1
torchvision==0.24.1
scikit-learn>=1.7.0
onnx>=1.20.0
onnxruntime>=1.23.0; python_version < "3.13"
PyYAML>=6.0
tqdm>=4.67.1
psutil>=5.9.0
memory-profiler>=0.60.0

# Qualité / tests
flake8>=6.0.0
black>=23.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pre-commit>=3.0.0

# Sécurité et monitoring
safety>=2.3.0
bandit>=1.7.0
locust>=2.15.0
